feature(groups): adjusting api description #PES-311
