+++ src/modules/group/utils/group.policies.ts
@@ -1,7 +1,7 @@
 import { BadRequestException, Injectable } from '@nestjs/common';
 import { Dictionary } from 'odyssey-dictionary';
 import { Group } from '../../../core/database/entities';
-import { UpsertGroupDTO } from '../dtos/upsert-group.dto';
+import { UpsertGroupDTO } from '../dto/upsert-group.dto';
 
 @Injectable()
 export class GroupPolicies {
